W-CERT-main/w-cert-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
gspread==6.0.0
oauth2client==4.1.3
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.4
